requirements.txt
langchain
langchain-community
langchain-openai
langchain-chroma
pymupdf
chromadb
streamlit
ragas
python-dotenv
openai
datasets
pandas
